requirements.txt
apturl==0.5.2
atomicwrites==1.1.5
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cmapy==0.6.6
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.30
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.6.0
fonttools==4.33.3
future==0.18.2
Glances==3.2.6.4
h5py==3.6.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.19.2
importlib-metadata==1.5.0
jdcal==1.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
networkx==2.8.1
numexpr==2.7.1
numpy==1.23.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==3.0.3
packaging==21.3
pandas==0.25.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==0.13.0
protobuf==3.6.1
psutil==5.9.1
py==1.8.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.3.0
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzed==3.7
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.19.2
scipy==1.8.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
systemd-python==234
tables==3.6.1
tifffile==2022.5.4
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchsummary==1.5.1
torchvision==0.12.0+cu113
typing-extensions==4.2.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
youtube-dl==2020.3.24
zipp==1.0.0
